explorer: Prettify program logs for transactions page
