+++ explorer/src/components/instruction/UnknownDetailsCard.tsx
@@ -5,6 +5,8 @@ import {
   ParsedInstruction,
 } from "@solana/web3.js";
 import { InstructionCard } from "./InstructionCard";
+import { programLabel } from "utils/tx";
+import { useCluster } from "providers/cluster";
 
 export function UnknownDetailsCard({
   ix,
@@ -19,12 +21,15 @@ export function UnknownDetailsCard({
   innerCards?: JSX.Element[];
   childIndex?: number;
 }) {
+  const { cluster } = useCluster();
+  const programName =
+    programLabel(ix.programId.toBase58(), cluster) || "Unknown Program";
   return (
     <InstructionCard
       ix={ix}
       index={index}
       result={result}
-      title="Unknown"
+      title={`${programName}: Unknown Instruction`}
       innerCards={innerCards}
       childIndex={childIndex}
       defaultRaw
