+++ explorer/src/components/instruction/mango/GenericMngoAccountDetailsCard.tsx
@@ -27,7 +27,7 @@ export function GenericMngoAccountDetailsCard(props: {
       ix={ix}
       index={index}
       result={result}
-      title={"Mango: " + title}
+      title={"Mango Program: " + title}
       innerCards={innerCards}
       childIndex={childIndex}
     >
