+++ explorer/src/components/instruction/token/TokenDetailsCard.tsx
@@ -40,7 +40,7 @@ export function TokenDetailsCard(props: DetailsProps) {
     const parsed = create(props.ix.parsed, ParsedInfo);
     const { type: rawType, info } = parsed;
     const type = create(rawType, TokenInstructionType);
-    const title = `Token: ${IX_TITLES[type]}`;
+    const title = `Token Program: ${IX_TITLES[type]}`;
     const created = create(info, IX_STRUCTS[type] as any);
     return <TokenInstruction title={title} info={created} {...props} />;
   } catch (err) {
