+++ explorer/src/pages/inspector/SimulatorCard.tsx
@@ -2,19 +2,8 @@ import React from "react";
 import bs58 from "bs58";
 import { Connection, Message, Transaction } from "@solana/web3.js";
 import { useCluster } from "providers/cluster";
-import { TableCardBody } from "components/common/TableCardBody";
-import { programLabel } from "utils/tx";
-
-type LogMessage = {
-  text: string;
-  prefix: string;
-  style: "muted" | "info" | "success" | "warning";
-};
-
-type InstructionLogs = {
-  logs: LogMessage[];
-  failed: boolean;
-};
+import { InstructionLogs, prettyProgramLogs } from "utils/program-logs";
+import { ProgramLogsCardBody } from "components/ProgramLogsCardBody";
 
 const DEFAULT_SIGNATURE = bs58.encode(Buffer.alloc(64).fill(0));
 
@@ -78,46 +67,7 @@ export function SimulatorCard({ message }: { message: Message }) {
           Retry
         </button>
       </div>
-      <TableCardBody>
-        {message.instructions.map((ix, index) => {
-          const programId = message.accountKeys[ix.programIdIndex];
-          const programName =
-            programLabel(programId.toBase58(), cluster) || "Unknown";
-          const programLogs: InstructionLogs | undefined = logs[index];
-
-          let badgeColor = "white";
-          if (programLogs) {
-            badgeColor = programLogs.failed ? "warning" : "success";
-          }
-
-          return (
-            <tr key={index}>
-              <td>
-                <div className="d-flex align-items-center">
-                  <span className={`badge badge-soft-${badgeColor} mr-2`}>
-                    #{index + 1}
-                  </span>
-                  {programName} Instruction
-                </div>
-                {programLogs && (
-                  <div className="d-flex align-items-start flex-column text-monospace p-2 font-size-sm">
-                    {programLogs.logs.map((log, key) => {
-                      return (
-                        <span key={key}>
-                          <span className="text-muted">{log.prefix}</span>
-                          <span className={`text-${log.style}`}>
-                            {log.text}
-                          </span>
-                        </span>
-                      );
-                    })}
-                  </div>
-                )}
-              </td>
-            </tr>
-          );
-        })}
-      </TableCardBody>
+      <ProgramLogsCardBody message={message} logs={logs} cluster={cluster} />
     </div>
   );
 }
@@ -152,124 +102,8 @@ function useSimulator(message: Message) {
         // Simulate without signers to skip signer verification
         const resp = await connection.simulateTransaction(tx);
 
-        let depth = 0;
-        let instructionLogs: InstructionLogs[] = [];
-        const prefixBuilder = (depth: number) => {
-          const prefix = new Array(depth - 1).fill("\u00A0\u00A0").join("");
-          return prefix + "> ";
-        };
-
-        let instructionError;
-        const responseLogs = resp.value.logs;
-        const responseErr = resp.value.err;
-        if (!responseLogs) {
-          if (resp.value.err) throw new Error(JSON.stringify(responseErr));
-          throw new Error("No logs detected");
-        } else if (responseErr) {
-          if (typeof responseErr !== "string") {
-            let ixError = (responseErr as any)["InstructionError"];
-            const [index, message] = ixError;
-            if (typeof message === "string") {
-              instructionError = { index, message };
-            }
-          } else {
-            throw new Error(responseErr);
-          }
-        }
-
-        responseLogs.forEach((log) => {
-          if (log.startsWith("Program log:")) {
-            instructionLogs[instructionLogs.length - 1].logs.push({
-              prefix: prefixBuilder(depth),
-              text: log,
-              style: "muted",
-            });
-          } else {
-            const regex = /Program (\w*) invoke \[(\d)\]/g;
-            const matches = [...log.matchAll(regex)];
-
-            if (matches.length > 0) {
-              const programAddress = matches[0][1];
-              const programName =
-                programLabel(programAddress, cluster) ||
-                `Unknown (${programAddress}) Program`;
-
-              if (depth === 0) {
-                instructionLogs.push({
-                  logs: [],
-                  failed: false,
-                });
-              } else {
-                instructionLogs[instructionLogs.length - 1].logs.push({
-                  prefix: prefixBuilder(depth),
-                  style: "info",
-                  text: `Invoking ${programName}`,
-                });
-              }
-
-              depth++;
-            } else if (log.includes("success")) {
-              instructionLogs[instructionLogs.length - 1].logs.push({
-                prefix: prefixBuilder(depth),
-                style: "success",
-                text: `Program returned success`,
-              });
-              depth--;
-            } else if (log.includes("failed")) {
-              const instructionLog =
-                instructionLogs[instructionLogs.length - 1];
-              if (!instructionLog.failed) {
-                instructionLog.failed = true;
-                instructionLog.logs.push({
-                  prefix: prefixBuilder(depth),
-                  style: "warning",
-                  text: `Program returned error: ${log.slice(
-                    log.indexOf(": ") + 2
-                  )}`,
-                });
-              }
-              depth--;
-            } else {
-              if (depth === 0) {
-                instructionLogs.push({
-                  logs: [],
-                  failed: false,
-                });
-                depth++;
-              }
-              // system transactions don't start with "Program log:"
-              instructionLogs[instructionLogs.length - 1].logs.push({
-                prefix: prefixBuilder(depth),
-                text: log,
-                style: "muted",
-              });
-            }
-          }
-        });
-
-        // If the instruction's simulation returned an error without any logs then add an empty log entry for Runtime error
-        // For example BpfUpgradableLoader fails without returning any logs for Upgrade instruction with buffer that doesn't exist
-        if (instructionError && instructionLogs.length === 0) {
-          instructionLogs.push({
-            logs: [],
-            failed: true,
-          });
-        }
-
-        if (
-          instructionError &&
-          instructionError.index === instructionLogs.length - 1
-        ) {
-          const failedIx = instructionLogs[instructionError.index];
-          failedIx.failed = true;
-          failedIx.logs.push({
-            prefix: prefixBuilder(1),
-            text: `Runtime error: ${instructionError.message}`,
-            style: "warning",
-          });
-        }
-
-        setLogs(instructionLogs);
+        // Prettify logs
+        setLogs(prettyProgramLogs(resp.value.logs, resp.value.err, cluster));
       } catch (err) {
         console.error(err);
         setLogs(null);
